--- tests/Tests.FeatureManagement.AspNetCore/Features.cs
@@ -1,11 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-//
-namespace Tests.FeatureManagement.AspNetCore
-{
-    enum Features
-    {
-        ConditionalFeature,
-        ConditionalFeature2
-    }
-}
