--- tests/Tests.FeatureManagement.AspNetCore/Pages/_ViewImports.cshtml
@@ -1,3 +1,0 @@
-﻿@using Tests.FeatureManagement.AspNetCore
-@namespace Tests.FeatureManagement.AspNetCore.Pages
-@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
