--- tests/Tests.FeatureManagement.AspNetCore/appsettings.json
@@ -1,28 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Warning"
-    }
-  },
-  "AllowedHosts": "*",
-
-  "FeatureManagement": {
-    "ConditionalFeature": {
-      "EnabledFor": [
-        {
-          "Name": "Test",
-          "Parameters": {
-            "P1": "V1"
-          }
-        }
-      ]
-    },
-    "ConditionalFeature2": {
-      "EnabledFor": [
-        {
-          "Name": "Test"
-        }
-      ]
-    }
-  }
-}
