+++ tests/Tests.FeatureManagement/Pages/_ViewImports.cshtml
@@ -1,0 +1,3 @@
+﻿@using Tests.FeatureManagement
+@namespace Tests.FeatureManagement.Pages
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
